abstract agent policies from qmix to be used in BC
