+++ jaxmarl/policies/__init__.py
@@ -1,0 +1,1 @@
+from .policies import ScannedRNN, AgentMLP, AgentHyperMLP, AgentRNN, AgentHyperRNN, HyperNetwork
